+++ tests/unit/Badoo/WithRestrictedConstantsPHP71TestClass.php
@@ -10,30 +10,30 @@ class WithRestrictedConstantsPHP71TestClass
     private const PRIVATE_VALUE = 1;
     protected const PROTECTED_VALUE = 11;
 
-    public static function getPrivateValue()
+    public static function getPrivateValue() : int
     {
         return self::PRIVATE_VALUE;
     }
 
-    public function getObjectPrivateValue()
-    {
-        return self::PRIVATE_VALUE;
-    }
-
-    public static function getProtectedValue()
+    public static function getSelfProtectedValue() : int
     {
         return self::PROTECTED_VALUE;
     }
 
-    public function getObjectProtectedValue()
+    public static function getStaticProtectedValue() : int
     {
-        return self::PROTECTED_VALUE;
+        return static::PROTECTED_VALUE;
+    }
+
+    public function getThisObjectProtectedValue() : int
+    {
+        return $this::PROTECTED_VALUE;
     }
 }
 
 class WithWrongPrivateConstantAccessPHP71TestClass
 {
-    public static function getPrivateValue()
+    public static function getPrivateValue() : int
     {
         return WithRestrictedConstantsPHP71TestClass::PRIVATE_VALUE;
     }
@@ -41,7 +41,7 @@ class WithWrongPrivateConstantAccessPHP71TestClass
 
 class WithWrongProtectedConstantAccessPHP71TestClass
 {
-    public static function getProtectedValue()
+    public static function getProtectedValue() : int
     {
         return WithRestrictedConstantsPHP71TestClass::PROTECTED_VALUE;
     }
@@ -49,23 +49,23 @@ class WithWrongProtectedConstantAccessPHP71TestClass
 
 class WithRestrictedConstantsChildPHP71TestClass extends WithRestrictedConstantsPHP71TestClass
 {
-    public static function getParentPrivateValue()
+    public static function getParentPrivateValue() : int
     {
         return parent::PRIVATE_VALUE;
     }
 
-    public static function getParentProtectedValue()
+    public static function getParentProtectedValue() : int
     {
         return parent::PROTECTED_VALUE;
     }
 }
 
-function getPrivateValue()
+function getPrivateValue() : int
 {
     return WithRestrictedConstantsPHP71TestClass::PRIVATE_VALUE;
 }
 
-function getProtectedValue()
+function getProtectedValue() : int
 {
     return WithRestrictedConstantsPHP71TestClass::PROTECTED_VALUE;
 }